Enable Unit tests for protobuf tool and package refactor
