+++ stdlib/ballerina-grpc/src/main/java/org/ballerinalang/net/grpc/nativeimpl/servicestub/AbstractExecute.java
@@ -31,7 +31,7 @@ import org.ballerinalang.util.codegen.PackageInfo;
 import org.ballerinalang.util.codegen.StructInfo;
 
 import static org.ballerinalang.net.grpc.MessageConstants.CONNECTOR_ERROR;
-import static org.ballerinalang.net.grpc.MessageConstants.PROTOCOL_PACKAGE_GRPC;
+import static org.ballerinalang.net.grpc.MessageConstants.PROTOCOL_STRUCT_PACKAGE_GRPC;
 
 /**
  * {@code AbstractExecute} is the Execute action implementation of the gRPC Connector.
@@ -39,12 +39,12 @@ import static org.ballerinalang.net.grpc.MessageConstants.PROTOCOL_PACKAGE_GRPC;
  * @since 1.0.0
  */
 public abstract class AbstractExecute extends BlockingNativeCallableUnit {
-
+    
     /**
      * Returns corresponding Ballerina type for the proto buffer type.
      *
      * @param protoType Protocol buffer type
-     * @param context Ballerina Context
+     * @param context   Ballerina Context
      * @return .
      */
     BType getBalType(String protoType, Context context) {
@@ -67,9 +67,9 @@ public abstract class AbstractExecute extends BlockingNativeCallableUnit {
             return context.getProgramFile().getEntryPackage().getStructInfo(protoType).getType();
         }
     }
-
+    
     MethodDescriptor.MethodType getMethodType(Descriptors.MethodDescriptor
-                                                                    methodDescriptor) throws GrpcClientException {
+                                                      methodDescriptor) throws GrpcClientException {
         if (methodDescriptor == null) {
             throw new GrpcClientException("Error while processing method type. Method descriptor cannot be null.");
         }
@@ -86,15 +86,15 @@ public abstract class AbstractExecute extends BlockingNativeCallableUnit {
             return MethodDescriptor.MethodType.UNKNOWN;
         }
     }
-
+    
     BStruct createStruct(Context context, String structName) {
         PackageInfo httpPackageInfo = context.getProgramFile()
-                .getPackageInfo(PROTOCOL_PACKAGE_GRPC);
+                .getPackageInfo(PROTOCOL_STRUCT_PACKAGE_GRPC);
         StructInfo structInfo = httpPackageInfo.getStructInfo(structName);
         BStructType structType = structInfo.getType();
         return new BStruct(structType);
     }
-
+    
     void notifyErrorReply(Context context, String errorMessage) {
         BStruct outboundError = createStruct(context, CONNECTOR_ERROR);
         outboundError.setStringField(0, errorMessage);
