+++ tests/ballerina-test/pom.xml
@@ -88,6 +88,14 @@
             <artifactId>mockito-all</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.ballerinalang</groupId>
+            <artifactId>ballerina-grpc</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.ballerinalang</groupId>
+            <artifactId>protobuf-ballerina</artifactId>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
