+++ tests/ballerina-test/src/test/resources/protoFiles/helloWorldClientStreaming.proto
